backend/requirements.txt
# Flask相关
Flask>=3.0.0
flask-cors>=4.0.0

# 数据库
pymongo>=4.6.0
mongoengine>=0.27.0

# 环境变量
python-dotenv>=1.0.0

# HTTP请求
requests>=2.31.0

# AI API客户端
huggingface-hub>=0.20.0

# 图片处理
Pillow>=10.0.0

# 数据验证
pydantic>=2.5.0

# 日志
python-json-logger>=2.0.7

# 工具
python-dateutil>=2.8.2

